Move password forms into diffrent namespace
